refactor: 💡 move input to components dir
